Refactored so electron.js uses the same reducer as renderers.
